Set up permissions for Genres and applied checks in routes

+++ routes/genres.js
@@ -3,6 +3,7 @@ const bodyParser = require('koa-bodyparser');
 const model = require('../models/genres');
 const auth = require('../controllers/auth');
 const {validateGenre, validateGenreUpdate} = require('../controllers/validation');
+const can = require('../permissions/genres');
 
 const router = Router({prefix: '/api/v1/genres'});
 
@@ -25,7 +26,7 @@ async function getAll(ctx) {
     }
 }
 
-  async function getById(ctx) {
+async function getById(ctx) {
     let id = ctx.params.id;
     let genre = await model.getById(id);
     if (genre.length) {
@@ -35,36 +36,69 @@ async function getAll(ctx) {
         ctx.status = 404;
     }
 }
-  
-  async function createGenre(ctx) {
+
+async function createGenre(ctx) {
     let body = ctx.request.body;
-    let result = await model.add(body);
-    if (result.affectedRows) {
-        ctx.body = {ID: result.insertId, created: true}
-        ctx.status = 201;
+    let permission = can.create(ctx.state.user);
+
+    if (!permission.granted) {
+        ctx.status = 403;
+    } else {
+        let result = await model.add(body);
+        if (result.affectedRows) {
+            ctx.body = {ID: result.insertId, created: true}
+            ctx.status = 201;
+        } else {
+            ctx.status = 404;
+        }
+    }
+}
+
+async function updateGenre(ctx){
+    let id = ctx.params.id;
+    let result = await model.getById(id);
+
+    if (result.length) {
+        let data = result[0];
+        let permission = can.update(ctx.state.user);
+
+        if (!permission.granted) {
+            ctx.status = 403;
+        } else {
+            let data = ctx.request.body;
+            let result = await model.updateById(data, id);
+            if (result.affectedRows) {
+                ctx.body = {ID: id, updated: true};
+                ctx.status = 200;
+            } else {
+                ctx.status = 404;
+            }
+        }
     } else {
         ctx.status = 404;
     }
 }
-  
-  async function updateGenre(ctx){
+
+async function deleteGenre(ctx){
     let id = ctx.params.id;
-    let body = ctx.request.body;
-    let result = await model.updateById(body, id);
-    if (result.affectedRows) {
-        ctx.body = {ID: id, updated: true};
-        ctx.status = 200;
-    } else {
-        ctx.status = 404;
-    }
-}
-  
-  async function deleteGenre(ctx){
-    let id = ctx.params.id;
-    let result = await model.deleteById(id);
-    if (result.affectedRows) {
-        ctx.body = {ID: id, deleted: true};
-        ctx.status = 200;
+    let result = await model.getById(id);
+
+    if (result.length) {
+        let data = result[0];
+        let permission = can.delete(ctx.state.user);
+
+        if (!permission.granted) {
+            ctx.status = 403;
+        } else {
+
+            let result = await model.deleteById(id);
+            if (result.affectedRows) {
+                ctx.body = {ID: id, deleted: true};
+                ctx.status = 200;
+            } else {
+                ctx.status = 404;
+            }
+        }       
     } else {
         ctx.status = 404;
     }
